+++ libs/designsystem/list/src/list-item/devicetype.ts
@@ -1,0 +1,1 @@
+export type DeviceType = 'desktop' | 'mobile';
