+++ libs/designsystem/list/src/list-item/list-item.component.ts
@@ -2,17 +2,8 @@ import { Component, EventEmitter, Input, Output, TemplateRef } from '@angular/co
 import { PlatformService, ThemeColor } from '@kirbydesign/designsystem/helpers';
 import { BoundaryClass } from '../list-item-swipe/list-item-swipe.component';
 import { ListSwipeAction } from '../list-swipe-action.type';
-
-export type DeviceType = 'desktop' | 'mobile';
-export interface ListItem {
-  id: number;
-  title: string;
-  subTitle: string;
-  amount: string;
-  detail: number;
-  flagged: boolean;
-  color: 'default' | 'primary' | 'success' | 'warning' | 'danger' | 'light';
-}
+import { DeviceType } from './devicetype';
+import { ListItem } from './list-item';
 
 @Component({
   selector: 'kirby-list-item',
