+++ libs/designsystem/list/src/list-item/list-item.ts
@@ -1,0 +1,9 @@
+export interface ListItem {
+  id: number;
+  title: string;
+  subTitle: string;
+  amount: string;
+  detail: number;
+  flagged: boolean;
+  color: 'default' | 'primary' | 'success' | 'warning' | 'danger' | 'light';
+}
