+++ libs/designsystem/list/src/list-item/pipes/has-actions/has-actions.pipe.ts
@@ -1,6 +1,7 @@
 import { Pipe, PipeTransform } from '@angular/core';
 import { ItemSlidingSide } from '@kirbydesign/designsystem/item-sliding';
-import { DeviceType, ListItem } from '../../list-item.component';
+import { DeviceType } from '../../devicetype';
+import { ListItem } from '../../list-item';
 
 export type ListItemAction = {
   isDisabled?: ((item: ListItem) => boolean) | boolean;
